requirements.txt
numpy==1.24.4
scipy==1.10.1
scikit-image==0.21.0
opencv-python==4.7.0.72
pillow==9.4.0
diffusers==0.25.1
transformers==4.41.1
accelerate
matplotlib==3.7.4
tqdm==4.64.1
config==0.5.1
einops==0.7.0
onnxruntime==1.16.2
basicsr
av
fvcore
cloudpickle
omegaconf
pycocotools
huggingface_hub==0.25.2
gradio
peft
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart
gdown
